Add option to disable initConnInfo
initConnInfo is making some assumption about how the virtual machines
are configured and installed. It is trying to add a usernet interface
as net1 and this can collide with a real net1 interface. It also
adds a PCI device at a hardcoded address which can also collide with
other configured hardware. And it assumes that the PROXMOX provisioner
is also used to tear down this interface after installation.